Added init.pp in the repo
Here I have change the content attribute. The line content => "PATH=/java/bin\n" should be changed to content => "PATH=\$PATH:/java/bin\n" since we want to append /java/bin to the environment variable $PATH, and not completely replace the current content in $PATH.
On top of this, files in the /etc/profile.d directory should only be editable by the root user. In order to do this we'll need to change the mode of the file to avoid giving others permission to write the file. In other words, the mode should be 0644 not 0646. So it has been done.